+++ frontend/src/lib/constants/canister-actor.constants.ts
@@ -1,5 +1,6 @@
 import { FETCH_ROOT_KEY, HOST } from "$lib/constants/environment.constants";
-import type { CanisterActorParams } from "$lib/types/canister";
+
+import type { CanisterActorParams } from "$lib/types/worker";
 
 export const ACTOR_PARAMS: Pick<CanisterActorParams, "host" | "fetchRootKey"> =
   {
