+++ frontend/src/lib/types/canister.ts
@@ -1,5 +1,4 @@
 import type { CanisterDetails } from "$lib/canisters/ic-management/ic-management.canister.types";
-import type { Identity } from "@dfinity/agent";
 import type { Principal } from "@dfinity/principal";
 
 export type CanisterSyncStatus = "syncing" | "synced" | "error";
@@ -14,10 +13,3 @@ export interface CanisterSync {
 }
 
 export type CanisterId = Principal;
-
-export interface CanisterActorParams {
-  canisterId: CanisterId | string;
-  identity: Identity;
-  host: string;
-  fetchRootKey: boolean;
-}
